allDjangoProjects/redwood/requirements.txt
backports.csv==1.0.7
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.8
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.2.0
docopt==0.6.2
et-xmlfile==1.0.1
fake-useragent==0.1.11
idna==2.8
jdcal==1.4
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
pipreqs==0.4.9
PyMySQL==0.9.3
pytz==2019.1
PyYAML==5.1
requests==2.21.0
soupsieve==1.9.1
sqlparse==0.3.0
tablib==0.13.0
urllib3==1.24.2
xlrd==1.2.0
xlwt==1.3.0
yarg==0.1.9
